fix: ty type-checker — 61 diagnostics opgelost
Echte bugs:
- tools/plot.py: return type tuple[str, go.Figure] → go.Figure | None (error paths retourneerden None)
- core/errors.py: list[tuple[type, str]] → str | None (BadRequestError heeft geen bericht)
- data/dashboard.py: return type dict[int, int] → dict[int, float] (round() geeft float terug)
- routes/chat.py: assert current_task is not None na _task_busy check (ty-narrowing)
- agent/run.py: ongebruikte ty: ignore comment verwijderd

Tests beter getypeerd:
- test_snippet.py: assert snippet is not None vóór in-checks
- test_tools_analysis.py: assert isinstance(result, str) vóór str-operaties
- test_tools_plot.py: assert fig is not None vóór fig.data-access
- test_instellingen.py: assert result is not None vóór dict-subscript

+++ data/dashboard.py
@@ -298,7 +298,7 @@ def _mbo_peer_eerstejaars(
     code_col: str,
     inst_col: str,
     peer_codes: frozenset,
-) -> tuple[dict[int, int], dict[str, dict[int, int]]]:
+) -> tuple[dict[int, float], dict[str, dict[int, int]]]:
     """Return (benchmark_eerstejaars, peers_eerstejaars) for MBO peers."""
     if df_instromende is None or df_instromende.empty:
         return {}, {}
@@ -1296,7 +1296,7 @@ def _rendement_ho(instelling: str) -> dict | None:
 
     hu_dipl = df_dipl[df_dipl[inst_col].str.lower() == instelling.lower()]
     dipl_series = hu_dipl.groupby("DIPLOMAJAAR")["AANTAL_GEDIPLOMEERDEN"].sum() if not hu_dipl.empty else pd.Series(dtype=int)
-    dipl = {int(k): int(v) for k, v in dipl_series.items()}
+    dipl = {int(k): int(v) for k, v in dipl_series.items()}  # ty: ignore[invalid-argument-type]
 
     cohorten = _pseudo_cohorten(instroom, dipl)
     result["pseudo_cohorten"] = cohorten
